- continued refactoring batteryStatus files - added more unittests
